Refactor: 일부 파일 리팩토링 및 네이밍 변경, CopyToClipBoard의 alert를 toast로 변경 #51
