+++ utils/copyToClipBoard.ts
@@ -1,8 +1,10 @@
+import Toast from './notification';
+
 export default async function copyToClipBoard(url: string) {
   try {
     await navigator.clipboard.writeText(url);
-    alert('복사 성공!');
+    Toast.success('복사에 성공했습니다.');
   } catch {
-    alert('복사 실패');
+    Toast.error('복사에 실패했습니다. 다시 시도해주세요.');
   }
 }
